Hackfix for fzf not working on gvim in windows.
Digging through the sourcecode, it turns out that the plugin does invoke the find command and redirects its output into a temp file.
This temp file is afterwards read in and processed. An analysis of the tempfiles shows that when run from gvim contains exactly
the selected file. On the invocation from cygwin terminal, however, the file contains one additional blank line at the beginning.
Hence, the hack is to insert that blank line after reading in the temp file if it consists of less than two lines.

This for sure does not cover all usecases and probably breaks stuff...
...but i is fixing the important usecase of opening a single fzf file.